# backend/src/apis/bonus-computation-api.js
const BonusComputation = require('../models/BonusComputation.model');
const Salesman = require('../models/Salesman.model');
const orangeHrmService = require('../services/orange-hrm-service');

exports.create = async function(req, res) {
    try {
        // create bonusComputation
        const bonusComputation = new BonusComputation({
            sid: req.body.sid,
            year: req.body.year,
            value: req.body.value,
            performanceRecords: req.body.performanceRecords,
            salesOrders: req.body.salesOrders,
            remarks: req.body.remarks
        });
        // save bonusComputation in database
        const data = await bonusComputation.save();
        // post bonusComputation to OrangeHRM
        const employeeId = await Salesman.findById(req.body.sid, 'employeeId');
        await orangeHrmService.postBonusComputation(employeeId.employeeId, req.body.year, req.body.value);
        res.send(data);
    }
    catch (err){
        res.status(500).send({
            message: err || 'There was a problem creating the bonus computation!'
        })
    }
};

exports.findBySidAndYear = async function(req, res) {
    try {
        const bonusComputation = await BonusComputation.find({sid: req.params.sid, year: req.params.year});
        if(!bonusComputation) {
            return res.status(404).send({message: "Bonus Computation not found!"});
        }
        res.send(bonusComputation);
    }
    catch(err) {
        return res.status(400).send({message: err.message});
    }
};