# backend/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Swagger of Alexander Tewes, Mario Hönighausen, Jan Frericks",
    "version": "1.0"
  },
  "produces": ["application/json"],
  "host": "localhost:8080",
  "basePath": "/api",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/performance-record/:sid:year": {
      "get": {
        "tags": [
          "Performance Record"
        ],
        "summary": "Get one performanceRecord by SalesmanId and Year",
        "operationId": "findBySidandYear",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Id and Year of salesOrder object that needs to be fetched",
            "schema": {
              "$ref": "#/definitions/id_year"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successful"
          },
          "401": {
            "description": "unauthorized"
          }
        }
      }
    },
...
    "/bonus-computation/:sid/:year": {
      "get": {
        "tags": [
          "Bonus Computation"
        ],
        "summary": "Get bonus computation by sid and year",
        "operationId": "findBySidAndYear",
        "consumes": [
          "application/json"
        ],
        "parameters": [{
          "name": "body",
          "description": "id and year of performanceRecord that needs to be fetched",
          "in": "body",
          "schema": {
            "$ref": "#/definitions/id_year"
          }
        }],
        "responses": {
          "201": {
            "description": "successful"
          },
          "401": {
            "description": "unauthorized"
          }
        }
      }
    },
    "/salesorder": {
      "post": {
        "tags": [
          "Salesorder"
        ],
        "summary": "Creates salesOrder",
        "operationId": "findBySidAndYear",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "salesOrder object that needs to be added",
            "schema": {
              "$ref": "#/definitions/Salesorder"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successful"
          },
          "401": {
            "description": "unauthorized"
          }
        }
      }
    },
    "/salesorder/:sid/:year": {
      "get": {
        "tags": [
          "Salesorder"
        ],
        "summary": "Get salesOrder by sid and year",
        "operationId": "findBySidAndYear",
        "consumes": [
          "application/json"
        ],
        "parameters": [{
          "name": "body",
          "description": "id and year of salesOrder that needs to be fetched",
          "in": "body",
          "schema": {
            "$ref": "#/definitions/id_year"
          }
        }],
        "responses": {
          "201": {
            "description": "successful"
          },
          "401": {
            "description": "unauthorized"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Salesorder": {
      "properties": {
        "sid": {
          "type": "number",
          "format": "int64",
          "example": "01"
        },
        "year":{
          "type": "number",
          "format": "int64",
          "example": 2000
        },
        "product": {
          "type": "string",
          "example": "Hoover"
        },
        "customerName": {
          "type": "string",
          "example": "Telekom"
        },
        "clientRanking": {
          "type": "number",
          "format": "int64",
          "example": "15"
        },
        "items": {
          "type": "number",
          "format": "int64",
          "example": "32"
        },
        "bonus": {
          "type": "number",
          "format": "int64",
          "example": "150"
        }
      }
    },
...
    "id_year":{
      "properties": {
        "sid": {
          "type": "number",
          "description": "sid of salesman",
          "example": 1
        },
        "year": {
          "type": "number",
          "description": "year",
          "example": 2000
        }
      }
    }
    }
  }